requirements.txt
asgiref==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect2==0.5.0
dj-database-url==0.5.0
Django==3.1.5
django-appconf==1.0.5
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-solo==2.0.0
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
Pillow==8.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2020.5
reportlab==3.6.8
requests==2.27.1
sqlparse==0.4.1
svglib==1.2.1
tinycss2==1.1.1
urllib3==1.26.8
webencodings==0.5.1
whitenoise==6.0.0
